services/ingestor_generator/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
example==0.1.dev0
idna==3.7
lxml==5.2.2
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.31.0
typing_extensions==4.11.0
-e git+https://github.com/josefdeutsch/udacity-capstone-python.git@f2897c52fa94545ebbd9659704beb39cac95f45e#egg=Udacity_capstone_python
urllib3==2.2.1
